Made sync more robust
Port selfupdate is tried several times until it succeeds.
Notice that the verify its success it tests for the presence of
a known port (xdrfile). Apparently, using the return status
of the `port selfupdate` command is not sufficient.

At the same time, I changed the default sync more back to
rsync which is faster.